Solve Day 3 Part 2.
The trick was to just use the "find the first largest number" algorithm repeatedly, but only finding the largest number with the right number of values left.

ie: If we find the first largest number with at least 11 numbers left at index 2, then search index 3 after to find the first largest number after that with at least 10 numbers left, etc.